refactor: implement BPT in terms of OZ ERC20 and ERC20Permit
