+++ pkg/pool-utils/contracts/BalancerPoolToken.sol
@@ -15,9 +15,8 @@
 pragma solidity ^0.7.0;
 
 import "@balancer-labs/v2-solidity-utils/contracts/math/Math.sol";
-import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/IERC20.sol";
-import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/IERC20Permit.sol";
-import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/EIP712.sol";
+import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/ERC20.sol";
+import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/ERC20Permit.sol";
 
 /**
  * @title Highly opinionated token implementation
@@ -32,186 +31,52 @@ import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/EIP712.sol";
  *   without first setting allowance
  * - Emits 'Approval' events whenever allowance is changed by `transferFrom`
  */
-contract BalancerPoolToken is IERC20, IERC20Permit, EIP712 {
-    using Math for uint256;
+contract BalancerPoolToken is ERC20, ERC20Permit {
+    constructor(string memory tokenName, string memory tokenSymbol)
+        ERC20(tokenName, tokenSymbol)
+        ERC20Permit(tokenName)
+    {}
 
-    // State variables
-
-    uint8 private constant _DECIMALS = 18;
-
-    mapping(address => uint256) private _balance;
-    mapping(address => mapping(address => uint256)) private _allowance;
-    uint256 private _totalSupply;
-
-    string private _name;
-    string private _symbol;
-
-    mapping(address => uint256) private _nonces;
-
-    // solhint-disable-next-line var-name-mixedcase
-    bytes32 private immutable _PERMIT_TYPE_HASH = keccak256(
-        "Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)"
-    );
-
-    // Function declarations
-
-    constructor(string memory tokenName, string memory tokenSymbol) EIP712(tokenName, "1") {
-        _name = tokenName;
-        _symbol = tokenSymbol;
-    }
-
-    // External functions
-
-    function allowance(address owner, address spender) external view override returns (uint256) {
-        return _allowance[owner][spender];
-    }
-
-    function balanceOf(address account) external view override returns (uint256) {
-        return _balance[account];
-    }
-
-    function approve(address spender, uint256 amount) external override returns (bool) {
-        _setAllowance(msg.sender, spender, amount);
-
-        return true;
-    }
-
-    function increaseApproval(address spender, uint256 amount) external returns (bool) {
-        _setAllowance(msg.sender, spender, _allowance[msg.sender][spender].add(amount));
-
-        return true;
-    }
-
-    function decreaseApproval(address spender, uint256 amount) external returns (bool) {
-        uint256 currentAllowance = _allowance[msg.sender][spender];
-
-        if (amount >= currentAllowance) {
-            _setAllowance(msg.sender, spender, 0);
-        } else {
-            _setAllowance(msg.sender, spender, currentAllowance.sub(amount));
-        }
-
-        return true;
-    }
-
-    function transfer(address recipient, uint256 amount) external override returns (bool) {
-        _move(msg.sender, recipient, amount);
-
-        return true;
-    }
+    // Public functions
 
     function transferFrom(
         address sender,
         address recipient,
         uint256 amount
-    ) external override returns (bool) {
-        uint256 currentAllowance = _allowance[sender][msg.sender];
+    ) public override returns (bool) {
+        uint256 currentAllowance = allowance(sender, msg.sender);
         _require(msg.sender == sender || currentAllowance >= amount, Errors.INSUFFICIENT_ALLOWANCE);
 
-        _move(sender, recipient, amount);
+        _transfer(sender, recipient, amount);
 
         if (msg.sender != sender && currentAllowance != uint256(-1)) {
             // Because of the previous require, we know that if msg.sender != sender then currentAllowance >= amount
-            _setAllowance(sender, msg.sender, currentAllowance - amount);
+            _approve(sender, msg.sender, currentAllowance - amount);
         }
 
         return true;
     }
 
-    function permit(
-        address owner,
-        address spender,
-        uint256 value,
-        uint256 deadline,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
-    ) public virtual override {
-        // solhint-disable-next-line not-rely-on-time
-        _require(block.timestamp <= deadline, Errors.EXPIRED_PERMIT);
+    function decreaseAllowance(address spender, uint256 amount) public override returns (bool) {
+        uint256 currentAllowance = allowance(msg.sender, spender);
 
-        uint256 nonce = _nonces[owner];
+        if (amount >= currentAllowance) {
+            _approve(msg.sender, spender, 0);
+        } else {
+            // No risk of underflow due to if condition
+            _approve(msg.sender, spender, currentAllowance - amount);
+        }
 
-        bytes32 structHash = keccak256(abi.encode(_PERMIT_TYPE_HASH, owner, spender, value, nonce, deadline));
-
-        bytes32 hash = _hashTypedDataV4(structHash);
-
-        address signer = ecrecover(hash, v, r, s);
-        _require((signer != address(0)) && (signer == owner), Errors.INVALID_SIGNATURE);
-
-        _nonces[owner] = nonce + 1;
-        _setAllowance(owner, spender, value);
-    }
-
-    // Public functions
-
-    function name() public view returns (string memory) {
-        return _name;
-    }
-
-    function symbol() public view returns (string memory) {
-        return _symbol;
-    }
-
-    function decimals() public pure returns (uint8) {
-        return _DECIMALS;
-    }
-
-    function totalSupply() public view override returns (uint256) {
-        return _totalSupply;
-    }
-
-    function nonces(address owner) external view override returns (uint256) {
-        return _nonces[owner];
-    }
-
-    // solhint-disable-next-line func-name-mixedcase
-    function DOMAIN_SEPARATOR() external view override returns (bytes32) {
-        return _domainSeparatorV4();
+        return true;
     }
 
     // Internal functions
 
     function _mintPoolTokens(address recipient, uint256 amount) internal {
-        _balance[recipient] = _balance[recipient].add(amount);
-        _totalSupply = _totalSupply.add(amount);
-        emit Transfer(address(0), recipient, amount);
+        _mint(recipient, amount);
     }
 
     function _burnPoolTokens(address sender, uint256 amount) internal {
-        uint256 currentBalance = _balance[sender];
-        _require(currentBalance >= amount, Errors.INSUFFICIENT_BALANCE);
-
-        _balance[sender] = currentBalance - amount;
-        _totalSupply = _totalSupply.sub(amount);
-        emit Transfer(sender, address(0), amount);
-    }
-
-    function _move(
-        address sender,
-        address recipient,
-        uint256 amount
-    ) internal {
-        uint256 currentBalance = _balance[sender];
-        _require(currentBalance >= amount, Errors.INSUFFICIENT_BALANCE);
-        // Prohibit transfers to the zero address to avoid confusion with the
-        // Transfer event emitted by `_burnPoolTokens`
-        _require(recipient != address(0), Errors.ERC20_TRANSFER_TO_ZERO_ADDRESS);
-
-        _balance[sender] = currentBalance - amount;
-        _balance[recipient] = _balance[recipient].add(amount);
-
-        emit Transfer(sender, recipient, amount);
-    }
-
-    // Private functions
-
-    function _setAllowance(
-        address owner,
-        address spender,
-        uint256 amount
-    ) private {
-        _allowance[owner][spender] = amount;
-        emit Approval(owner, spender, amount);
+        _burn(sender, amount);
     }
 }
